Make asset analysis methods free top-level functions in asset.py, fix typos, simplify methods to find keywords
